requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
backcall==0.2.0
bleach==3.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.4
cssselect2==0.4.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-appconf==1.0.2
django-compressor==2.2
django-formtools==2.2
django-tools==0.48.3
django-wkhtmltopdf==3.3.0
django-xhtml2pdf==0.0.4
future==0.18.2
html5lib==1.1
icdiff==1.9.1
ipython==7.0.0
ipython-genutils==0.2.0
jedi==0.18.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.0.0
pprintpp==0.4.0
prompt-toolkit==2.0.10
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dotenv==0.8.2
pytz==2018.3
rcssmin==1.0.6
reportlab==3.5.59
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
traitlets==5.0.5
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.5
